LC-count-number-of-trapezoids-i [Runtime: 153 ms; Memory: 219.2 MB]

+++ array/count-number-of-trapezoids-i.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int countTrapezoids(vector<vector<int>>& points) {
-        unordered_map<int,int>level;
+        unordered_map<int,long long>level;
 int const mod=1e9+7;
         for(auto p:points){
             int i=p[1];
@@ -11,7 +11,7 @@ int const mod=1e9+7;
         long long sum=0;
         for(auto [p,t]:level){
             if(t<2)continue;
-            int c=((t*(t-1))/2);
+            long long c=((t*(t-1))/2);
             ans+=sum*c;
             ans%=mod;
             sum+=c;
